+++ packages/uniforms/test/bridges/index.js
@@ -7,6 +7,22 @@ describe('createSchemaBridge', () => {
         expect(createSchemaBridge).to.be.a('function');
     });
 
+    it('accepts a Bridge instance', () => {
+        const bridge = {
+            getError () {},
+            getErrorMessage () {},
+            getErrorMessages () {},
+            getField () {},
+            getInitialValue () {},
+            getProps () {},
+            getSubfields () {},
+            getType () {},
+            getValidator () {}
+        };
+
+        expect(createSchemaBridge(bridge)).to.be.equal(bridge);
+    });
+
     it('throws on unrecognised schema', () => {
         expect(createSchemaBridge).to.throw(/Unrecognised schema: /);
     });